+++ pkg/serverless/aws/arn.go
@@ -27,6 +27,15 @@ const (
 	RegionEnvVar       = "AWS_REGION"
 	functionNameEnvVar = "AWS_LAMBDA_FUNCTION_NAME"
 	qualifierEnvVar    = "AWS_LAMBDA_FUNCTION_VERSION"
+
+	traceOriginMetadataKey   = "_dd.origin"
+	traceOriginMetadataValue = "lambda"
+	computeStatsKey          = "_dd.compute_stats"
+	computeStatsValue        = "1"
+	functionARNKey           = "function_arn"
+	functionNameKey          = "functionname"
+	regionKey                = "region"
+	awsAccountKey            = "aws_account"
 )
 
 type persistedState struct {
@@ -221,6 +230,18 @@ func GetARNTags() []string {
 	}
 	tags = append(tags, fmt.Sprintf("resource:%s", resource))
 
+	return tags
+}
+
+// BuildGlobalTagsMap returns tags associated with the given ARN
+func BuildGlobalTagsMap(functionARN string, functionName string, region string, awsAccountID string) map[string]string {
+	tags := make(map[string]string)
+	tags[traceOriginMetadataKey] = traceOriginMetadataValue
+	tags[computeStatsKey] = computeStatsValue
+	tags[functionARNKey] = functionARN
+	tags[functionNameKey] = functionName
+	tags[regionKey] = region
+	tags[awsAccountKey] = awsAccountID
 	return tags
 }
